move pbnj server to own cmd directory
moving into its own subdirectory allows users to go install it as
well. such as `go install github.com/tinkerbell/pbnj/cmd/pbnj` which
will install the command as `pbnj` since `go install` uses the directory
as the name for the command.
