+++ cmd/pbnj/server.go
@@ -1,4 +1,4 @@
-package cmd
+package main
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 	grpc_validator "github.com/grpc-ecosystem/go-grpc-middleware/validator"
 	"github.com/packethost/pkg/log/logr"
 	"github.com/spf13/cobra"
-	"github.com/tinkerbell/pbnj/cmd/zaplog"
+	"github.com/tinkerbell/pbnj/cmd/pbnj/zaplog"
 	"github.com/tinkerbell/pbnj/server/grpcsvr"
 	"goa.design/goa/grpc/middleware"
 	"google.golang.org/grpc"
